refactor(controller/source_folders): extract entry update helpers
